+++ app/src/main/java/com/kyhsgeekcode/disassembler/MainActivity.java
@@ -17,7 +17,6 @@ import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.res.Resources;
-import android.database.Cursor;
 import android.graphics.Color;
 import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
@@ -25,9 +24,6 @@ import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Environment;
-import android.provider.DocumentsContract;
-import android.provider.MediaStore;
 import android.util.Log;
 import android.util.LongSparseArray;
 import android.util.SparseArray;
@@ -68,24 +64,20 @@ import com.github.chrisbanes.photoview.PhotoView;
 import com.kyhsgeekcode.disassembler.Calc.Calculator;
 import com.kyhsgeekcode.disassembler.FileTabFactory.FileTabContentFactory;
 import com.kyhsgeekcode.disassembler.Utils.Olly.UddTag;
+import com.kyhsgeekcode.disassembler.Utils.Util;
 import com.stericson.RootTools.RootTools;
 
 import org.apache.commons.io.FilenameUtils;
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
 import java.io.DataInputStream;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -97,7 +89,6 @@ import java.util.concurrent.LinkedBlockingQueue;
 import java.util.regex.Pattern;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
-import java.util.zip.ZipOutputStream;
 
 import at.pollaknet.api.facile.Facile;
 import at.pollaknet.api.facile.FacileReflector;
@@ -161,11 +152,9 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
     Spinner spinnerArch;
     TabHost tabHost;
     LinearLayout tab1, tab2;
-    //FileTabContentFactory factory = new FileTabContentFactory(this);
 
     ///////////////////////////////////////////////////UI manager////////////////////////////////////////////
     HexManager hexManager = new HexManager();
-
 
     Queue<Runnable> toDoAfterPermQueue = new LinkedBlockingQueue<>();
 
@@ -245,10 +234,6 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
     private TextView tvAnalRes;
     private ImageView ivAnalCount;
 
-    private long instantEntry;
-    private Capstone cs;
-    private String EXTRA_NOTIFICATION_ID;
-    private String ACTION_SNOOZE;
     private ProjectManager projectManager;
     private ProjectManager.Project currentProject;
     private ListView lvSymbols;
@@ -259,11 +244,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
     private View.OnClickListener leftListener = new View.OnClickListener() {
         public void onClick(View v) {
             ColumnSetting cs = (ColumnSetting) v.getTag();
-			/*String hint=(String) ((Button)v).getHint();
-			hint=hint.substring(1,hint.length()-1);
-			Log.v(TAG,"Hint="+hint);
-			String [] parsed=hint.split(", ",0);
-			Log.v(TAG,Arrays.toString(parsed));*/
+
             columnSetting = cs;
             setShowAddress(cs.showAddress/*Boolean.valueOf(parsed[1]*/);///*v.getTag(CustomDialog.TAGAddress)*/);
             setShowLabel(cs.showLabel/*Boolean.valueOf(parsed[0]*/);///*v.getTag(CustomDialog.TAGLabel)*/);
@@ -736,10 +717,10 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
                 ShowDetail();
                 break;
             case R.id.btnSaveDisasm:
-                ExportDisasm();
+                //ExportDisasm();
                 break;
             case R.id.btnSaveDetails:
-                SaveDetail();
+                //SaveDetail();
                 break;
             case R.id.mainBTFinishSetup: {
                 if (parsedFile == null) {
@@ -842,7 +823,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
                 tabHost.setCurrentTab(TAB_EXPORT);
                 return;
             }
-        }
+        }/*
         if (shouldSave && currentProject == null) {
             ShowYesNoCancelDialog(this, "Save project?", "",
                     new DialogInterface.OnClickListener() {
@@ -869,7 +850,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
                         public void onClick(DialogInterface p1, int p2) {
                         }
                     });
-        } else
+        } else*/
             super.onBackPressed();
     }
 
@@ -883,9 +864,6 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
 		 catch (Exception e)
 		 {}*/
         Finalize();
-        if (cs != null)
-            ;//cs.close();
-        cs = null;
     }
 
     @Override
@@ -1049,12 +1027,12 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
             case R.id.save: {
                 //if(currentProject==null)
                 {
-                    ExportDisasm(this::SaveDetail);
+                    //ExportDisasm(this::SaveDetail);
                 }
                 break;
             }
             case R.id.export: {
-                ExportDisasm(new Runnable() {
+                /*ExportDisasm(new Runnable() {
                     @Override
                     public void run() {
                         SaveDetail(new Runnable() {
@@ -1064,7 +1042,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
                             }
                         });
                     }
-                });
+                });*/
 
                 break;
             }
@@ -1343,303 +1321,6 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
         showChooser();/*File*/
     }
 
-    /////////////////////////////////////////////Export - Output//////////////////////////////////
-    public void ExportDisasm() {
-        ExportDisasm(null);
-    }
-
-    private void ExportDisasm(final Runnable runnable) {
-        requestAppPermissions(this);
-        if (fpath == null || "".compareToIgnoreCase(fpath) == 0) {
-            AlertSelFile();
-            return;
-        }
-        Toast.makeText(this, "Sorry, not stable yet", Toast.LENGTH_SHORT).show();
-        if (true)
-            return;
-        if (currentProject == null) {
-            final EditText etName = new EditText(this);
-            ShowEditDialog(getString(R.string.newProject), getString(R.string.enterNewProjName), etName, getString(R.string.ok), new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface p1, int p2) {
-
-                    String projn = etName.getText().toString();
-                    SaveDisasmNewProject(projn, runnable);
-                }
-            }, getString(R.string.cancel), new DialogInterface.OnClickListener() {
-
-                @Override
-                public void onClick(DialogInterface p1, int p2) {
-                }
-            });
-        } else {
-            ShowExportOptions(runnable);
-        }
-
-    }
-
-    //FIXME, TODO
-    private void ExportDisasmSub(int mode) {
-        Log.v(TAG, "Saving disassembly");
-        if (mode == 0)//Raw mode
-        {
-            SaveDisasmRaw();
-            return;
-        }
-        if (mode == 4)//Database mode
-        {
-            SaveDisasm(currentProject.getDisasmDb());
-            return;
-        }
-        File dir = new File(ProjectManager.RootFile, currentProject.name + "/");
-        Log.d(TAG, "dirpath=" + dir.getAbsolutePath());
-        File file = new File(dir, "Disassembly_" + new Date(System.currentTimeMillis()).toString() + (mode == 3 ? ".json" : ".txt"));
-        Log.d(TAG, "filepath=" + file.getAbsolutePath());
-        dir.mkdirs();
-        try {
-            file.createNewFile();
-        } catch (IOException e) {
-            Log.e(TAG, "", e);
-            Toast.makeText(this, R.string.failSaveFile, Toast.LENGTH_SHORT).show();
-        }
-        //Editable et=etDetails.getText();
-        try {
-            FileOutputStream fos = new FileOutputStream(file);
-            try {
-                StringBuilder sb = new StringBuilder();
-                ArrayList<ListViewItem>/*ListViewItem[]*/ items = new ArrayList<>();
-                //items.addAll(adapter.itemList());
-                for (ListViewItem lvi : items) {
-                    switch (mode) {
-                        case 1:
-                            sb.append(lvi.address);
-                            sb.append("\t");
-                            sb.append(lvi.bytes);
-                            sb.append("\t");
-                            sb.append(lvi.instruction);
-                            sb.append(" ");
-                            sb.append(lvi.operands);
-                            sb.append("\t");
-                            sb.append(lvi.comments);
-                            break;
-                        case 2:
-                            sb.append(lvi.address);
-                            sb.append(":");
-                            sb.append(lvi.instruction);
-                            sb.append(" ");
-                            sb.append(lvi.operands);
-                            sb.append("  ;");
-                            sb.append(lvi.comments);
-                            break;
-                        case 3:
-                            sb.append(lvi.toString());
-                    }
-                    sb.append(System.lineSeparator());
-                }
-                fos.write(sb.toString().getBytes());
-            } catch (IOException e) {
-                AlertError("", e);
-                return;
-            }
-        } catch (FileNotFoundException e) {
-            AlertError("", e);
-        }
-        AlertSaveSuccess(file);
-    }
-
-    private void SaveDisasmRaw() {
-        File dir = new File(ProjectManager.RootFile, currentProject.name + "/");
-        Log.d(TAG, "dirpath=" + dir.getAbsolutePath());
-        File file = new File(dir, "Disassembly.raw");
-        Log.d(TAG, "filepath=" + file.getAbsolutePath());
-        dir.mkdirs();
-        try {
-            file.createNewFile();
-        } catch (IOException e) {
-            Log.e(TAG, "", e);
-            Toast.makeText(this, R.string.failSaveFile, Toast.LENGTH_SHORT).show();
-        }
-        try {
-            FileOutputStream fos = new FileOutputStream(file);
-            ObjectOutputStream oos = new ObjectOutputStream(fos);
-            oos.writeObject(disasmResults);
-            oos.close();
-        } catch (IOException e) {
-            AlertError(getString(R.string.failSaveFile), e);
-            return;
-        }
-        AlertSaveSuccess(file);
-    }
-
-    private void SaveDetail() {
-        SaveDetail(null);
-    }
-
-    private void SaveDetail(final Runnable runnable) {
-        requestAppPermissions(this);
-        if (fpath == null || "".compareToIgnoreCase(fpath) == 0) {
-            AlertSelFile();
-            return;
-        }
-        if (currentProject == null) {
-            final EditText etName = new EditText(this);
-            ShowEditDialog(getString(R.string.newProject), getString(R.string.enterNewProjName), etName, getString(R.string.ok), new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface p1, int p2) {
-
-                    String projn = etName.getText().toString();
-                    SaveDetailNewProject(projn);
-                    if (runnable != null)
-                        runnable.run();
-                }
-            }, getString(R.string.cancel), new DialogInterface.OnClickListener() {
-
-                @Override
-                public void onClick(DialogInterface p1, int p2) {
-
-                }
-            });
-        } else {
-            try {
-                SaveDetailSub(currentProject);
-                if (runnable != null)
-                    runnable.run();
-            } catch (IOException e) {
-                AlertError(getString(R.string.failSaveFile), e);
-            }
-        }
-
-        //SaveDetailOld();
-    }
-
-    private void SaveDetail(File dir, File file) {
-        dir.mkdirs();
-        try {
-            file.createNewFile();
-        } catch (IOException e) {
-            Log.e(TAG, "", e);
-            Toast.makeText(this, R.string.failSaveFile, Toast.LENGTH_SHORT).show();
-        }
-
-        try {
-            FileOutputStream fos = new FileOutputStream(file);
-            try {
-                fos.write(parsedFile.toString().getBytes());
-            } catch (IOException e) {
-                Log.e(TAG, "", e);
-            }
-        } catch (FileNotFoundException e) {
-            Log.e(TAG, "", e);
-        }
-
-        AlertSaveSuccess(file);
-    }
-
-    private void SaveDetailNewProject(String projn) {
-
-        try {
-            ProjectManager.Project proj = projectManager.newProject(projn, fpath);
-            proj.Open(false);
-            db = new DatabaseHelper(this, ProjectManager.createPath(proj.name) + "disasm.db");
-            SaveDetailSub(proj);
-        } catch (IOException e) {
-            AlertError(R.string.failCreateProject, e);
-        }
-    }
-
-    private void SaveDetailSub(ProjectManager.Project proj) throws IOException {
-        File detailF = proj.getDetailFile();
-        if (detailF == null)
-            throw new IOException("Failed to create detail File");
-        currentProject = proj;
-        detailF.createNewFile();
-        SaveDetail(new File(ProjectManager.Path), detailF);
-        proj.Save();
-    }
-
-    private void SaveDisasmNewProject(String projn) {
-        SaveDisasmNewProject(projn, null);
-    }
-
-    private void SaveDisasmNewProject(String projn, Runnable runnable) {
-        try {
-            ProjectManager.Project proj = projectManager.newProject(projn, fpath);
-            currentProject = proj;
-            proj.Open(false);
-            db = new DatabaseHelper(this, ProjectManager.createPath(proj.name) + "disasm.db");
-            ShowExportOptions(runnable);
-            proj.Save();
-
-        } catch (IOException e) {
-            AlertError(getString(R.string.failCreateProject), e);
-        }
-    }
-
-    private void ShowExportOptions() {
-        ShowExportOptions(null);
-    }
-
-    private void ShowExportOptions(final Runnable runnable) {
-        final List<String> ListItems = new ArrayList<>();
-        ListItems.add("Raw(Fast,Reloadable)");
-        ListItems.add("Classic(Addr bytes inst op comment)");
-        ListItems.add("Simple(Addr: inst op; comment");
-        ListItems.add("Json");
-        ListItems.add("Database(.db, reloadable)");
-        ShowSelDialog(this, ListItems, getString(R.string.export_as), new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int pos) {
-                //String selectedText = items[pos].toString();
-                dialog.dismiss();
-                final ProgressDialog dialog2 = showProgressDialog(getString(R.string.saving));
-                ExportDisasmSub(pos);
-                if (runnable != null)
-                    runnable.run();
-                dialog2.dismiss();
-            }
-        });
-    }
-
-    private void createZip() {
-        File targetFile;
-        try {
-            File projFolder = new File(ProjectManager.RootFile, currentProject.name + "/");
-            FileOutputStream fos = new FileOutputStream(targetFile = new File(ProjectManager.RootFile, currentProject.name + ".zip"));
-            ZipOutputStream zos = new ZipOutputStream(fos);
-            File[] targets = projFolder.listFiles();
-            byte[] buf = new byte[4096];
-            int readlen;
-            for (File file : targets) {
-                Log.v(TAG, "writing " + file.getName());
-                ZipEntry ze = new ZipEntry(file.getName());
-                zos.putNextEntry(ze);
-                FileInputStream fis = new FileInputStream(file);
-                while ((readlen = fis.read(buf, 0, 4096)) > 0)
-                    zos.write(buf, 0, readlen);
-                zos.closeEntry();
-                fis.close();
-            }
-            zos.close();
-            fos.close();
-        } catch (Exception e) {
-            AlertError(R.string.fail_exportzip, e);
-            targetFile = null;
-        }
-        if (targetFile != null)
-            AlertSaveSuccess(targetFile);
-    }
-
-    private void SaveDisasm(DatabaseHelper disasmF) {
-        new SaveDBAsync().execute(disasmF);
-    }
-
-    private void SaveDetailOld() {
-        Log.v(TAG, "Saving details");
-        File dir = new File(Environment.getExternalStorageDirectory().getPath() + "disasm/");
-        File file = new File(dir, new File(fpath).getName() + "_" + new Date(System.currentTimeMillis()).toString() + ".details.txt");
-        SaveDetail(dir, file);
-    }
-
-    ////////////////////////////////////////////End Export - Output/////////////////////////////////////////
     //////////////////////////////////////////////Projects////////////////////////////////////////////////////////////////////////
     @Override
     public void onOpen(ProjectManager.Project proj) {
@@ -1948,14 +1629,14 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
         try {
 
             InputStream is = getContentResolver().openInputStream(uri);
-            if (HandleZipFIle(getRealPathFromURI(uri), is)) {
+            if (HandleZipFIle(Util.getRealPathFromURI(this, uri), is)) {
                 return;
             }
-            if (HandleUddFile(getRealPathFromURI(uri), is)) {
+            if (HandleUddFile(Util.getRealPathFromURI(this, uri), is)) {
                 return;
             }
             //ByteArrayOutputStream bis=new ByteArrayOutputStream();
-            setFilecontent(Utils.getBytes(is));
+            setFilecontent(Util.getBytes(is));
 
             tmpfile.createNewFile();
             FileOutputStream fos = new FileOutputStream(tmpfile);
@@ -1972,7 +1653,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
                     }
                     try {
                         RootTools.copyFile(uri.getPath(), tmpfile.getPath(), false, false);
-                        setFilecontent(Utils.getBytes(new FileInputStream(tmpfile)));
+                        setFilecontent(Util.getBytes(new FileInputStream(tmpfile)));
                         setFpath(tmpfile.getAbsolutePath());//uri.getPath();
                         AfterReadFully(tmpfile);
                         return;
@@ -2019,7 +1700,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
             etFilename.setText(file.getAbsolutePath());
             long fsize = file.length();
             //int index = 0;
-            setFilecontent(Utils.getBytes(in)/*new byte[(int) fsize]*/);
+            setFilecontent(Util.getBytes(in)/*new byte[(int) fsize]*/);
             /*
             int len= 0;
             byte[] b = new byte[1024];
@@ -2045,7 +1726,7 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
                     }
                     try {
                         RootTools.copyFile(path, tmpfile.getPath(), false, false);
-                        setFilecontent(Utils.getBytes(new FileInputStream(tmpfile)));
+                        setFilecontent(Util.getBytes(new FileInputStream(tmpfile)));
                         setFpath(tmpfile.getAbsolutePath());//uri.getPath();
                         AfterReadFully(tmpfile);
                         return;
@@ -2288,35 +1969,6 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
     }
 
     ////////////////////////////////////////////Data Conversion//////////////////////////////////
-
-
-
-    private String getRealPathFromURI(Uri uri) {
-        String filePath;
-        filePath = uri.getPath();
-        //경로에 /storage가 들어가면 real file path로 판단
-        if (filePath.startsWith("/storage"))
-            return filePath;
-        String wholeID = DocumentsContract.getDocumentId(uri);
-        //wholeID는 파일명이 abc.zip이라면 /document/B5D7-1CE9:abc.zip와 같습니다.
-        // Split at colon, use second item in the array
-        String id = wholeID.split(":")[0];
-        //Log.e(TAG, "id = " + id);
-        String[] column = {MediaStore.Files.FileColumns.DATA};
-        //파일의 이름을 통해 where 조건식을 만듭니다.
-        String sel = MediaStore.Files.FileColumns.DATA + " LIKE '%" + id + "%'";
-        //External storage에 있는 파일의 DB를 접근하는 방법 입니다.
-        Cursor cursor = getContentResolver().query(MediaStore.Files.getContentUri("external"), column, sel, null, null);
-        //SQL문으로 표현하면 아래와 같이 되겠죠????
-        //SELECT _dtat FROM files WHERE _data LIKE '%selected file name%'
-        int columnIndex = cursor.getColumnIndex(column[0]);
-        if (cursor.moveToFirst()) {
-            filePath = cursor.getString(columnIndex);
-        }
-        cursor.close();
-        return filePath;
-    }
-
     private ProgressDialog showProgressDialog(String s) {
         ProgressDialog dialog = new ProgressDialog(this);
         dialog.setProgressStyle(ProgressDialog.STYLE_HORIZONTAL);
@@ -2397,108 +2049,6 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
 
     public static native int Open(int arch, int mode);
 
-    public static class Utils {
-        public static byte[] getBytes(InputStream is) throws IOException {
-
-            int len;
-            int size = 1024;
-            byte[] buf;
-
-            if (is instanceof ByteArrayInputStream) {
-                size = is.available();
-                buf = new byte[size];
-                len = is.read(buf, 0, size);
-            } else {
-                ByteArrayOutputStream bos = new ByteArrayOutputStream();
-                buf = new byte[size];
-                while ((len = is.read(buf, 0, size)) != -1)
-                    bos.write(buf, 0, len);
-                buf = bos.toByteArray();
-            }
-            is.close();
-            return buf;
-        }
-    }
-
-    class SaveDBAsync extends AsyncTask<DatabaseHelper, Integer, Void> {
-        String TAG = getClass().getSimpleName();
-        android.app.AlertDialog.Builder builder;
-        ProgressBar progress;
-
-        protected void onPreExecute() {
-            super.onPreExecute();
-            Log.d(TAG + " PreExceute", "On pre Exceute......");
-            progress = new ProgressBar(MainActivity.this);
-            progress.setIndeterminate(false);
-
-            builder = new android.app.AlertDialog.Builder(MainActivity.this);
-            builder.setTitle("Saving..").setView(progress);
-            builder.show();
-        }
-
-        protected Void doInBackground(DatabaseHelper... disasmF) {
-            Log.d(TAG + " DoINBackGround", "On doInBackground...");
-
-            int cnt = disasmF[0].getCount();
-            if (cnt == 0) {
-                int datasize = disasmResults.size();
-                for (int i = 0; i < datasize; ++i) {
-                    //disasmF[0].insert(disasmResults.get(i));
-                    publishProgress(i);
-                }
-            }
-            return null;
-        }
-
-        protected void onProgressUpdate(Integer... a) {
-            super.onProgressUpdate(a);
-            progress.setProgress(a[0]);
-            //Log.d(TAG + " onProgressUpdate", "You are in progress update ... " + a[0]);
-        }
-		/*
-		 protected void onPostExecute(Void result) {
-		 super.onPostExecute(result);
-		 //Log.d(TAG + " onPostExecute", "" + result);
-		 }
-		 */
-    }
-
-    class SaveDisasmAsync extends AsyncTask<Void, Integer, Void> {
-        //String TAG = getClass().getSimpleName();
-        android.app.AlertDialog.Builder builder;
-        ProgressBar progress;
-
-        protected void onPreExecute() {
-            super.onPreExecute();
-            Log.d(TAG + " PreExceute", "On pre Exceute......");
-            progress = new ProgressBar(MainActivity.this);
-            progress.setIndeterminate(false);
-
-            builder = new android.app.AlertDialog.Builder(MainActivity.this);
-            builder.setTitle("Saving..").setView(progress);
-            builder.show();
-        }
-
-        protected Void doInBackground(Void... list) {
-            Log.d(TAG + " DoINBkGnd", "On doInBackground...");
-            SaveDisasmRaw();
-            return null;
-        }
-
-        protected void onProgressUpdate(Integer... a) {
-            super.onProgressUpdate(a);
-            progress.setProgress(a[0]);
-            //Log.d(TAG + " onProgressUpdate", "You are in progress update ... " + a[0]);
-        }
-
-		/*
-		 protected void onPostExecute(Void result) {
-		 super.onPostExecute(result);
-		 //Log.d(TAG + " onPostExecute", "" + result);
-		 }
-		 */
-    }
-
     private class DrawerItemClickListener implements ListView.OnItemClickListener {
         @Override
         public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
@@ -2536,235 +2086,4 @@ public class MainActivity extends AppCompatActivity implements Button.OnClickLis
         textFileExts.add("il");
     }
 }
-	/*	OnCreate()
-	 vp = (ViewPager)findViewById(R.id.pager);
-	 Button btn_first = (Button)findViewById(R.id.btn_first);
-	 Button btn_second = (Button)findViewById(R.id.btn_second);
-	 Button btn_third = (Button)findViewById(R.id.btn_third);
-
-	 vp.setAdapter(new pagerAdapter(getSupportFragmentManager()));
-	 vp.setCurrentItem(0);
-
-	 btn_first.setOnClickListener(movePageListener);
-	 btn_first.setTag(0);
-	 btn_second.setOnClickListener(movePageListener);
-	 btn_second.setTag(1);
-	 btn_third.setOnClickListener(movePageListener);
-	 btn_third.setTag(2);*/
-// Adapter 생성
-// adapter = new DisasmListViewAdapter() ;
-	/*	ListViewItem item=new ListViewItem();
-	 item.setAddress("address");
-	 item.setBytes("bytes");
-	 item.setComments("comments");
-	 item.setCondition("condition");
-	 item.setInstruction("inst");
-	 item.setLabel("label");
-	 item.setOperands("operands");
-	 adapter.addItem(item);
-	 // 리스트뷰 참조 및 Adapter달기
-	 listview = (ListView) findViewById(R.id.lvDisassembly);
-	 listview.setAdapter(adapter);
-	 listview.setOnTouchListener(new ListView.OnTouchListener() {
-	 @Override
-	 public boolean onTouch(View v, MotionEvent event) {
-	 int action = event.getAction();
-	 switch (action) {
-	 case MotionEvent.ACTION_DOWN:
-	 // Disallow ScrollView to intercept touch events.
-	 v.getParent().requestDisallowInterceptTouchEvent(true);
-	 break;
-
-	 case MotionEvent.ACTION_UP:
-	 // Allow ScrollView to intercept touch events.
-	 v.getParent().requestDisallowInterceptTouchEvent(false);
-	 break;
-	 }
-
-	 // Handle ListView touch events.
-	 v.onTouchEvent(event);
-	 return true;
-	 }});
-	 // 위에서 생성한 listview에 클릭 이벤트 핸들러 정의.
-	 listview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-	 @Override
-	 public void onItemClick(AdapterView parent, View v, int position, long id) {
-	 // get item
-	 ListViewItem item = (ListViewItem) parent.getItemAtPosition(position) ;
-
-	 //String titleStr = item.getTitle() ;
-	 //String descStr = item.getDesc() ;
-	 //Drawable iconDrawable = item.getIcon() ;
-
-	 // TODO : use item data.
-	 }
-	 }) ;*/
-	/*
-	 PrintStackTrace to string
-	 ByteArrayOutputStream out = new ByteArrayOutputStream();
-	 PrintStream pinrtStream = new PrintStream(out);
-	 e.printStackTrace(pinrtStream);
-	 String stackTraceString = out.toString(); // 찍은 값을 가져오고.
-
-	 */
-	/*
-	 public void onWindowFocusChanged(boolean hasFocus) {
-	 // get content height
-	 int contentHeight = listview.getChildAt(0).getHeight();
-	 contentHeight*=listview.getChildCount();
-	 // set listview height
-	 LayoutParams lp = listview.getLayoutParams();
-	 lp.height = contentHeight;
-	 listview.setLayoutParams(lp);
-	 }
-	 */
-
-	/*    switch(id) {
-	 case R.id.menu_login:
-	 Toast.makeText(getApplicationContext(), "로그인 메뉴 클릭",
-	 Toast.LENGTH_SHORT).show();
-	 return true;
-	 case R.id.menu_logout:
-	 Toast.makeText(getApplicationContext(), "로그아웃 메뉴 클릭",
-	 Toast.LENGTH_SHORT).show();
-	 return true;
-	 case R.id.menu_a:
-	 Toast.makeText(getApplicationContext(), "다음",
-	 Toast.LENGTH_SHORT).show();
-	 return true;
-	 }*/
-	/*
-	 View.OnClickListener movePageListener = new View.OnClickListener()
-	 {
-	 @Override
-	 public void onClick(View v)
-	 {
-	 int tag = (int) v.getTag();
-	 vp.setCurrentItem(tag);
-	 }
-	 };
-
-	 private class pagerAdapter extends FragmentStatePagerAdapter
-	 {
-	 public pagerAdapter(android.support.v4.app.FragmentManager fm)
-	 {
-	 super(fm);
-	 }
-	 @Override
-	 public android.support.v4.app.Fragment getItem(int position)
-	 {
-	 switch(position)
-	 {
-	 case 0:
-	 return new OverviewFragment();
-	 case 1:
-	 return new OverviewFragment();
-	 case 2:
-	 return new OverviewFragment();
-	 default:
-	 return null;
-	 }
-	 }
-	 @Override
-	 public int getCount()
-	 {
-	 return 3;
-	 }
-	 }*/
-//details.setText("file format not recognized.");
-//	String result=sample.getText().toString();
-//Toast toast = Toast.makeText(myActivity, result, Toast.LENGTH_LONG);
-//toast.show();
-	/*PE pe=PEParser.parse(fpath);
-	 if (pe != null)
-	 {
-	 PESignature ps =pe.getSignature();
-	 if (ps == null || !ps.isValid())
-	 {
-	 //What is it?
-	 Toast.makeText(this, "The file seems that it is neither a valid Elf file or PE file!", Toast.LENGTH_SHORT).show();
-	 throw new IOException(e);
-	 }
-	 }
-	 else
-	 {
-	 //What is it?
-	 Toast.makeText(this, "The file seems that it is neither a valid Elf file or PE file!", Toast.LENGTH_SHORT).show();
-	 throw new IOException(e);
-	 }*/
-/*
-	 private void CreateDisasmTopRow(TableRow tbrow0)
-	 {
-	 TextView tv0 = new TextView(MainActivity.this);
-	 tv0.setText(" Address ");
-	 tv0.setTextColor(Color.BLACK);
-	 tbrow0.addView(tv0);
-	 TextView tv1 = new TextView(MainActivity.this);
-	 tv1.setText(" Label ");
-	 tv1.setTextColor(Color.BLACK);
-	 tbrow0.addView(tv1);
-	 TextView tv2 = new TextView(MainActivity.this);
-	 tv2.setText(" Bytes ");
-	 tv2.setTextColor(Color.BLACK);
-	 tbrow0.addView(tv2);
-	 TextView tv3 = new TextView(MainActivity.this);
-	 tv3.setText(" Inst ");
-	 tv3.setTextColor(Color.BLACK);
-	 tbrow0.addView(tv3);
-	 TextView tv4 = new TextView(MainActivity.this);
-	 tv4.setText(" Cond ");
-	 tv4.setTextColor(Color.BLACK);
-	 tbrow0.addView(tv4);
-	 TextView tv5 = new TextView(MainActivity.this);
-	 tv5.setText(" Operands ");
-	 tv5.setTextColor(Color.BLACK);
-	 tbrow0.addView(tv5);
-	 TextView tv6 = new TextView(MainActivity.this);
-	 tv6.setText(" Comment ");
-	 tv6.setTextColor(Color.BLACK);
-	 AdjustShow(tv0, tv1, tv2, tv3, tv4, tv5, tv6);
-	 tbrow0.addView(tv6);
-	 }
-	 */
-	 /*
-				 private String[] getAccounts() {
-				 Pattern emailPattern = Patterns.EMAIL_ADDRESS;
-				 Account[] accounts = AccountManager.get(MainActivity.this).getAccounts();
-				 if(accounts==null)
-				 {
-				 return new String[]{""};
-				 }
-				 ArrayList<String> accs=new ArrayList<>();
-				 for (Account account : accounts) {
-				 if (emailPattern.matcher(account.name).matches()) {
-				 String email = account.name;
-				 accs.add(email);
-				 //Log.d(TAG, "email : " + email);
-				 }
-				 }
-				 return accs.toArray(new String[accs.size()]);
-				 }*/
-/**
- * Swaps fragments in the main content view
- * <p>
- * private void selectItem(int position) {
- * //Project project=
- * // Create a new fragment and specify the planet to show based on position
- * /*Fragment fragment = new PlanetFragment();
- * Bundle args = new Bundle();
- * args.putInt(PlanetFragment.ARG_PLANET_NUMBER, position);
- * fragment.setArguments(args);
- * <p>
- * // Insert the fragment by replacing any existing fragment
- * FragmentManager fragmentManager = getFragmentManager();
- * fragmentManager.beginTransaction()
- * .replace(R.id.content_frame, fragment)
- * .commit();
- * <p>
- * // Highlight the selected item, update the title, and close the drawer
- * mDrawerList.setItemChecked(position, true);
- * setTitle(mPlanetTitles[position]);
- * mDrawerLayout.closeDrawer(mDrawerList);
- * }
- */
